refactor(ops): extract measure-classification & column-error helpers
New _measure_helpers.py (~280 lines):

- _extract_measure_metadata: read user-supplied measure forms
  (dict/Measure/raw) into a uniform tuple
- _is_calculated_measure / _matches_aggregation_pattern /
  _find_matching_measure: detect calc measures vs base measures and
  match aggregation patterns against known measure definitions
- _make_base_measure: wrap a callable/Deferred/AggregationExpr into a
  Measure
- _classify_measure: top-level dispatch — used externally by expr.py
- _build_json_definition: dim/measure JSON-export helper
- _format_column_error: friendly diagnostic for missing-column errors

_values.py's lazy import of _format_column_error switches to the new
home so the value-objects → core back-ref is gone. _core.py is now
3,283 lines (down from 3,551). 930 tests pass.
